// lib/schema.ts
export const PatientFormSchema = z.object({
  first_name: z.string().min(1),
  last_name: z.string().min(1),
  date_of_birth: z.date(), // ✅ match your form Date object
  gender: z.enum(["MALE", "FEMALE"]),
  phone: z.string(),
  email: z.string().email(),
  address: z.string(),
  marital_status: z.enum([
    "married",
    "single",
    "divorced",
    "widowed",
    "separated",
  ]),
  emergency_contact_name: z.string().optional(),
  emergency_contact_number: z.string().optional(),
  relation: z.enum(["mother", "father", "husband", "wife", "other"]),
  blood_group: z.string().optional(),
  allergies: z.string().optional(),
  medical_conditions: z.string().optional(),
  medical_history: z.string().optional(),
  insurance_provider: z.string().optional(),
  insurance_number: z.string().optional(),
  medical_consent: z.boolean().optional(),
  privacy_consent: z.boolean().optional(),
  service_consent: z.boolean().optional(),
  img: z.string().optional(),
});